I was curious what the following statement meant:
Dependencies don't store data, they just track the set of computations
to invalidate if something changes. Typically, a data value will be
accompanied by a Dependency object that tracks the computations that
depend on it.

What did they mean by *computations*. It's kind of eluded me up to now.
Essentially, it appears as though literally any function can be a *computation*
so long as that function has a dependancy statment in it that this *computation* will,
or function, will depend on.

Here you can see that a new function, getRandomNumber, has been declared.
It's used as the return from a helper. I've known helpers to be reactive, meaning
that if inside of a helper there is a dependency, then when that dependency
changes, the helper will re-run.

getRandomNumer just simply returns a random number out of it. But it also,
states that it has a dependency on the favoriteFoodDep dependancy.

So now, whenever a new favorite food is set, this will trigger the favoiriteFoodDep
to go into a changed state, and dependant *computations* will re-run.

Those computations are getFavoriteFood and randomNumber.

If we were to comment out the line:
favoriteFoodDep.depend();
from the getRandomNumber function, then when we set the favorite food we
would not get a new random number every time.

This all makes more sense to me, although I'm not entirely sure what I
typed above makes sense, or is 100% correct.