Set up a Tracker.autorun to observe it's behavior. The function given to autorun will run immediately once at least when the app loads in the client due to it asking for the favorite food, which has been made into a reactive variable. It will run regardless of what route you are on.
So if you visit http://localhost:3000 you will see it run once. This is
because the app has loaded, the favoriteFood variable has been set, and
a call to getFavoriteFood via the autorun has executed.

And if you visit http://localhost:3000/simpleexample you will see it run
once as well.

I would have thought that the second one would not have run though as the
variable had already been set. A little off in my mind. Curiouse what would
happen if you visited some other route, what would happen?

Flip some of the " to '

Remove the isClient wrapper as this file is now in the client folder and will
only run on the client now by convention.